Reconcile trybuild runner with main
Preserve the feature-tracking test orchestration while adopting main's
fallible snapshot helpers and conditional diagnostic normalization.

Move the serialized trybuild entrypoint into a private runner module so the
combined integration-test root stays within the repository file-size limit.
Record the rebase decisions in the active ExecPlan.
